fix: audio

+++ Game/Animation.cs
@@ -149,11 +149,11 @@ public class FrameData {
     public bool hasHit;
     public int Sprite_index { get; set; }
     public string Sound_index { get; set; }
-    public int DeltaX { get; set; }
-    public int DeltaY { get; set; }
+    public float DeltaX { get; set; }
+    public float DeltaY { get; set; }
     public List<GenericBox> Boxes { get; set; }
 
-    public FrameData(int sprite_index, int deltaX, int deltaY, List<GenericBox> boxes, string Sound_index = "", bool hasHit = true) {
+    public FrameData(int sprite_index, float deltaX, float deltaY, List<GenericBox> boxes, string Sound_index = "", bool hasHit = true) {
         this.Sprite_index = sprite_index;
         this.DeltaX = deltaX;
         this.DeltaY = deltaY;
